frontend/requirements.txt
aiohttp==3.7.4
async-timeout==3.0.1
attrs==21.4.0
chardet<4.0
discord.py==1.7.3
idna==3.3
multidict==6.0.2
typing_extensions==4.0.1
yarl==1.7.2
requests
